style: ReadQuestion -> CreateQuestion
